Rename file to match the name of the hook.
