+++ app/src/resources/labware_offsets/index.ts
@@ -1,1 +1,1 @@
-export * from './useNotifyLabwareOffsetsQuery'
+export * from './useNotifySearchLabwareOffsets'
